[REFACTOR] UserRoleEnum, User, UserService, UserController 클래스 경로 변경 및 import 수정
