+++ src/main/java/com/sparta/quizdemo/chat/entity/ChatRoom.java
@@ -1,6 +1,6 @@
 package com.sparta.quizdemo.chat.entity;
 
-import com.sparta.quizdemo.common.entity.User;
+import com.sparta.quizdemo.user.entity.User;
 import jakarta.persistence.*;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
