+++ src/main/java/com/sparta/quizdemo/chat/service/RedisSubscriber.java
@@ -4,11 +4,11 @@ import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.sparta.quizdemo.chat.entity.ChatMessage;
 import com.sparta.quizdemo.chat.repository.ChatMessageRepository;
-import com.sparta.quizdemo.common.entity.User;
+import com.sparta.quizdemo.user.entity.User;
 import com.sparta.quizdemo.sse.entity.NotificationType;
 import com.sparta.quizdemo.sse.service.NotificationService;
-import com.sparta.quizdemo.user.UserRepository;
-import com.sparta.quizdemo.user.UserService;
+import com.sparta.quizdemo.user.repository.UserRepository;
+import com.sparta.quizdemo.user.service.UserService;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.data.redis.connection.Message;
