+++ src/main/java/com/sparta/quizdemo/sse/dto/NotificationRequestDto.java
@@ -1,6 +1,6 @@
 package com.sparta.quizdemo.sse.dto;
 
-import com.sparta.quizdemo.common.entity.User;
+import com.sparta.quizdemo.user.entity.User;
 import com.sparta.quizdemo.sse.entity.NotificationType;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
