+++ src/main/java/com/sparta/quizdemo/user/entity/User.java
@@ -6,9 +6,6 @@ import com.sparta.quizdemo.common.entity.TimeStamped;
 import com.sparta.quizdemo.common.entity.UserRoleEnum;
 import com.sparta.quizdemo.order.entity.Order;
 import com.sparta.quizdemo.sse.entity.Notification;
-import com.sparta.quizdemo.user.SignupRequestDto;
-import com.sparta.quizdemo.user.UserRequestDto;
-import com.sparta.quizdemo.user.UserRoleEnum;
 import com.sparta.quizdemo.user.dto.SignupRequestDto;
 import com.sparta.quizdemo.user.dto.UserRequestDto;
 import jakarta.persistence.*;
@@ -16,6 +13,7 @@ import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
 
+import java.util.ArrayList;
 import java.util.List;
 
 @Entity
